refactor(cli): move args module under commands
