+++ crates/tsz-cli/src/lib.rs
@@ -4,7 +4,6 @@
 //! all CLI-specific modules (argument parsing, file discovery, config loading,
 //! compilation driver, watch mode, etc.).
 
-pub mod args;
 pub mod commands;
 pub use tsz::config;
 pub mod driver;
@@ -13,6 +12,7 @@ pub mod locale;
 pub mod project;
 pub mod reporting;
 pub mod watch;
+pub use commands::args;
 pub use commands::build;
 pub use commands::help;
 pub use project::fs;
